# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memorial website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteStats {
  totalVisitors: number;
  totalDuas: number;
  totalDhikr: number;
  totalIstighfar: number;
  totalSalawat: number;
  mostUsedDhikr: string;
  countries: string[];
}

export interface VisitInput {
  /** @nullable */
  country?: string | null;
}

export interface DhikrCounts {
  SubhanAllah: number;
  Alhamdulillah: number;
  AllahuAkbar: number;
  LaIlahaIllAllah: number;
  LaHawla: number;
  Astaghfirullah: number;
  SubhanAllahWaBiHamdihi: number;
  SubhanAllahAlAzeem: number;
  Salawat: number;
}

export interface DhikrIncrement {
  dhikrKey: string;
}

export interface Dua {
  id: number;
  text: string;
  createdAt: string;
}

export interface DuaInput {
  text: string;
}