# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Memorial website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dhikr/increment:
    post:
      operationId: incrementDhikr
      tags: [dhikr]
      summary: Increment a global dhikr counter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DhikrIncrement"
      responses:
        "200":
          description: Incremented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DhikrCounts"
...
components:
  schemas:
...
    DhikrCounts:
      type: object
      properties:
        SubhanAllah:
          type: number
        Alhamdulillah:
          type: number
        AllahuAkbar:
          type: number
        LaIlahaIllAllah:
          type: number
        LaHawla:
          type: number
        Astaghfirullah:
          type: number
        SubhanAllahWaBiHamdihi:
          type: number
        SubhanAllahAlAzeem:
          type: number
        Salawat:
          type: number
      required:
        - SubhanAllah
        - Alhamdulillah
        - AllahuAkbar
        - LaIlahaIllAllah
        - LaHawla
        - Astaghfirullah
        - SubhanAllahWaBiHamdihi
        - SubhanAllahAlAzeem
        - Salawat
    DhikrIncrement:
      type: object
      properties:
        dhikrKey:
          type: string
      required:
        - dhikrKey
...