# lib/api-zod/src/generated/types/dhikrCounts.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memorial website API
 * OpenAPI spec version: 0.1.0
 */

export interface DhikrCounts {
  SubhanAllah: number;
  Alhamdulillah: number;
  AllahuAkbar: number;
  LaIlahaIllAllah: number;
  LaHawla: number;
  Astaghfirullah: number;
  SubhanAllahWaBiHamdihi: number;
  SubhanAllahAlAzeem: number;
  Salawat: number;
}